+++ CHANGELOG.md
@@ -4,6 +4,13 @@
   - `AttributeType` no longer inherits from `BaseResource` (no API path), removing its `from_json` method and `relative_id` property
   - The type of `AttributeType`'s `attributes` property is now a `list` of `SubAttribute`s, which are identical to `Attribute`s except they lack an API path
   - The `Dataset` function `update_records` has been renamed `_update_records` as the convenience functions `upsert_records` and `delete_records` now exist.
+  - All files have been refactored:
+    * The `models` directory has been deleted, everything previously in it has been moved directly into the base directory
+    * `DatasetProfile` and `DatasetStatus` have been moved into the `dataset` directory
+    * `machine_learning_model.py` has been renamed `base_model.py`
+    * Attribute configurations have been moved to a subdirectory within `project`
+    * A `mastering` directory has been created with all mastering specific entities
+    * A `categorization` directory has been created with all categorization specific entities, including a `category` subdirectory
 
   **NEW FEATURES**
   - [#174](https://github.com/Datatamer/unify-client-python/issues/174) Get and create taxonomy categories